+++ backend/AnarchyChess.Ai/Models/MoveUndoState.cs
@@ -32,9 +32,9 @@ public unsafe struct MoveUndoState
 
     public required bool PrevIsWhiteToMove;
 
-    public required float PrevWhiteMaterialCount;
+    public required int PrevWhiteMaterialCount;
     public required int PrevWhiteKingCount;
-    public required float PrevBlackMaterialCount;
+    public required int PrevBlackMaterialCount;
     public required int PrevBlackKingCount;
 
     public void AddCapture(byte square, PieceType piece, BitPieceColor color)
